ISSUE #1407 - Set comment permissions from risk details
